requirements.txt
# === WEB FRAMEWORK ===
# FastAPI - Modern web framework for building the application
fastapi>=0.109.0

# Uvicorn - Web server that runs FastAPI
uvicorn[standard]>=0.27.0

# Gunicorn - Production web server (for deployment)
gunicorn>=21.2.0

# === DATABASE ===
# SQLAlchemy - Lets Python talk to databases (like an Excel file)
sqlalchemy>=2.0.36

# Async SQLite - Makes database faster
aiosqlite>=0.19.0

# === FILE HANDLING ===
# For handling file uploads (invoices, PDFs, etc.)
python-multipart>=0.0.6

# For reading dates in different formats
python-dateutil>=2.8.2

# For reading/writing files asynchronously
aiofiles>=23.2.1

# === EXCEL & PDF GENERATION ===
# OpenPyXL - Create Excel files for schedules
openpyxl>=3.1.2

# ReportLab - Create PDF files
reportlab>=4.0.8

# === IMAGE PROCESSING & VISION ===
# Pillow - Process images for AI vision
Pillow>=10.2.0

# pdf2image - Convert PDF attachments to images for AI vision
# Note: Requires poppler - see https://pdf2image.readthedocs.io/
pdf2image>=1.17.0

# === TEMPLATES ===
# Jinja2 - Generate HTML pages
jinja2>=3.1.3

# === EMAIL SERVICES ===
# Gmail API - For reading emails from Gmail
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.116.0

# Microsoft Graph API - For reading emails from Outlook
msal>=1.24.0

# === SECURITY ===
# Cryptography - For encrypting OAuth tokens in database
cryptography>=41.0.0
passlib>=1.7.4,<1.8.0
# bcrypt 4.1+ triggers noisy compatibility warnings in passlib 1.7.x
bcrypt>=4.0.1,<4.1.0

# === AI SERVICE ===
# HTTPX - Make API calls to OpenRouter (AI service)
httpx>=0.26.0

# Requests - Alternative HTTP library (used in some places)
requests>=2.31.0

# === CONFIGURATION ===
# Python-dotenv - Read .env file for configuration
python-dotenv>=1.0.0
